Improved logic for setting the target framework
The default value of the `Framework` parameter has been changed to an empty string, so that the latest available version is used as the default. Conditional logic has been added to `template.json` and `Program.cs` to set `TargetFramework` only if the `Framework` parameter is non-empty. The same changes have been applied to the WebAssembly template.